# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LongInvest",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/daily-data/batches": {
      "get": {
        "tags": [
          "daily-data"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_api_v1_daily_data_batches_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyBatchPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/daily-data/batches/{batch_id}/missing": {
      "get": {
        "tags": [
          "daily-data"
        ],
        "summary": "List Missing",
        "operationId": "list_missing_api_v1_daily_data_batches__batch_id__missing_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyMissingPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/daily-data/batches/{batch_id}/retry": {
      "post": {
        "tags": [
          "daily-data"
        ],
        "summary": "Retry Batch",
        "operationId": "retry_batch_api_v1_daily_data_batches__batch_id__retry_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/daily-bars/{symbol}/revisions": {
      "get": {
        "tags": [
          "daily-data"
        ],
        "summary": "List Revisions",
        "operationId": "list_revisions_api_v1_daily_bars__symbol__revisions_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyRevisionPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/daily-bars/{symbol}": {
      "get": {
        "tags": [
          "daily-data"
        ],
        "summary": "List Bars",
        "operationId": "list_bars_api_v1_daily_bars__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyBarPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DailyBarPageData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyBarRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "DailyBarPageData"
      },
      "DailyBarPageResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/DailyBarPageData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "DailyBarPageResponse"
      },
      "DailyBarRecord": {
        "properties": {
          "security_id": {
            "type": "string",
            "format": "uuid",
            "title": "Security Id"
          },
          "trade_date": {
            "type": "string",
            "format": "date",
            "title": "Trade Date"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "open": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Open"
          },
          "high": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "High"
          },
          "low": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Low"
          },
          "close": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Close"
          },
          "previous_close": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "volume": {
            "type": "integer",
            "title": "Volume"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "data_version": {
            "type": "integer",
            "title": "Data Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "security_id",
          "trade_date",
          "symbol",
          "open",
          "high",
          "low",
          "close",
          "previous_close",
          "volume",
          "amount",
          "source",
          "data_version",
          "created_at",
          "updated_at"
        ],
        "title": "DailyBarRecord"
      },
      "DailyBatchPageData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyBatchRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "DailyBatchPageData"
      },
      "DailyBatchPageResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/DailyBatchPageData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "DailyBatchPageResponse"
      },
      "DailyBatchRecord": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "trading_date": {
            "type": "string",
            "format": "date",
            "title": "Trading Date"
          },
          "universe_snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Universe Snapshot Id"
          },
          "parent_batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Batch Id"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "security_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Security Ids"
          },
          "known_corporate_action_symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Corporate Action Symbols"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "expected_count": {
            "type": "integer",
            "title": "Expected Count"
          },
          "fetched_count": {
            "type": "integer",
            "title": "Fetched Count"
          },
          "validated_count": {
            "type": "integer",
            "title": "Validated Count"
          },
          "committed_count": {
            "type": "integer",
            "title": "Committed Count"
          },
          "missing_count": {
            "type": "integer",
            "title": "Missing Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trading_date",
          "universe_snapshot_id",
          "parent_batch_id",
          "symbols",
          "security_ids",
          "known_corporate_action_symbols",
          "idempotency_key",
          "status",
          "expected_count",
          "fetched_count",
          "validated_count",
          "committed_count",
          "missing_count",
          "failed_count",
          "created_at",
          "started_at",
          "deadline_at",
          "completed_at"
        ],
        "title": "DailyBatchRecord"
      },
      "DailyJobData": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_type",
          "status"
        ],
        "title": "DailyJobData"
      },
      "DailyJobResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/DailyJobData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "DailyJobResponse"
      },
      "DailyMissingPageData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyMissingRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "DailyMissingPageData"
      },
      "DailyMissingPageResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/DailyMissingPageData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "DailyMissingPageResponse"
      },
      "DailyMissingRecord": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "security_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "explained": {
            "type": "boolean",
            "title": "Explained"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "security_id",
          "symbol",
          "reason",
          "error_code",
          "explained",
          "created_at"
        ],
        "title": "DailyMissingRecord"
      },
      "DailyRevisionPageData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyRevisionRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "DailyRevisionPageData"
      },
      "DailyRevisionPageResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/DailyRevisionPageData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "DailyRevisionPageResponse"
      },
      "DailyRevisionRecord": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "daily_bar_security_id": {
            "type": "string",
            "format": "uuid",
            "title": "Daily Bar Security Id"
          },
          "daily_bar_trade_date": {
            "type": "string",
            "format": "date",
            "title": "Daily Bar Trade Date"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "revision_no": {
            "type": "integer",
            "title": "Revision No"
          },
          "old_values": {
            "additionalProperties": true,
            "type": "object",
            "title": "Old Values"
          },
          "new_values": {
            "additionalProperties": true,
            "type": "object",
            "title": "New Values"
          },
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed Fields"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "daily_bar_security_id",
          "daily_bar_trade_date",
          "symbol",
          "revision_no",
          "old_values",
          "new_values",
          "changed_fields",
          "source",
          "reason",
          "created_at"
        ],
        "title": "DailyRevisionRecord"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Pagination": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total"
        ],
        "title": "Pagination"
      },
...
      "RetryRequest": {
        "properties": {
          "confirm": {
            "type": "boolean",
            "title": "Confirm"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirm",
          "reason"
        ],
        "title": "RetryRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}